sc-7295 gds user ui: refact addresses component
