+++ web/gds-user-ui/src/components/Addresses/AddressList.spec.tsx
@@ -1,14 +1,14 @@
 import userEvent from '@testing-library/user-event';
 import { dynamicActivate } from 'utils/i18nLoaderHelper';
 import { render, screen } from 'utils/test-utils';
-import Addresses from '.';
+import AddressList from './AddressList';
 
-describe('<Addresses />', () => {
+describe('<AddressList />', () => {
   beforeEach(() => {
     dynamicActivate('en');
   });
   it('should add address row', () => {
-    render(<Addresses />);
+    render(<AddressList />);
 
     const addAddress = screen.getByRole('button', { name: /add address/i });
